float128 does not exist in some architectures

+++ stingray/tests/test_gti.py
@@ -248,7 +248,7 @@ class TestGTI(object):
                         [1.16703386e+08, 1.16703418e+08], 
                         [1.16703418e+08, 1.16703450e+08], 
                         [1.16703450e+08, 1.16703482e+08], 
-                        [1.16703482e+08, 1.16703514e+08]], dtype=np.float128)
+                        [1.16703482e+08, 1.16703514e+08]])
         newg = join_equal_gti_boundaries(gti)
-        assert np.allclose(newg, np.array([[1.16703354e+08, 1.16703514e+08]], dtype=np.float128))
+        assert np.allclose(newg, np.array([[1.16703354e+08, 1.16703514e+08]]))
